+++ grafana-plugin/src/models/escalation_policy/escalation_policy.ts
@@ -34,26 +34,23 @@ export class EscalationPolicyStore extends BaseStore {
     this.path = '/escalation_policies/';
   }
 
-  @action
+  @action.bound
   async updateWebEscalationPolicyOptions() {
     const response = await makeRequest('/escalation_policies/escalation_options/', {});
-
     this.webEscalationChoices = response;
   }
 
-  @action
+  @action.bound
   async updateEscalationPolicyOptions() {
     const response = await makeRequest('/escalation_policies/', {
       method: 'OPTIONS',
     });
-
     this.escalationChoices = get(response, 'actions.POST', []);
   }
 
-  @action
+  @action.bound
   async updateNumMinutesInWindowOptions() {
     const response = await makeRequest('/escalation_policies/num_minutes_in_window_options/', {});
-
     this.numMinutesInWindowOptions = response;
   }