+++ grafana-plugin/src/network/index.ts
@@ -5,7 +5,6 @@ import qs from 'query-string';
 
 import FaroHelper from 'utils/faro';
 
-export const API_HOST = `${window.location.protocol}//${window.location.host}/`;
 export const API_PROXY_PREFIX = 'api/plugin-proxy/grafana-oncall-app';
 export const API_PATH_PREFIX = '/api/internal/v1';