+++ grafana-plugin/src/pages/integration/Integration.tsx
@@ -181,7 +181,7 @@ class Integration extends React.Component<IntegrationProps, IntegrationState> {
 
             <div className={cx('integration__heading-container')}>
               <PluginLink query={{ page: 'integrations', ...query }} className={cx('back-arrow')}>
-                <IconButton name="arrow-left" size="xl" />
+                <IconButton aria-label="Go Back" name="arrow-left" size="xl" />
               </PluginLink>
               <h2 className={cx('integration__name')}>
                 <Emoji text={alertReceiveChannel.verbal_name} />