+++ grafana-plugin/src/state/rootBaseStore/index.ts
@@ -19,6 +19,7 @@ import { GlobalSettingStore } from 'models/global_setting/global_setting';
 import { GrafanaTeamStore } from 'models/grafana_team/grafana_team';
 import { HeartbeatStore } from 'models/heartbeat/heartbeat';
 import { LabelStore } from 'models/label/label';
+import { LoaderStore } from 'models/loader/loader';
 import { OrganizationStore } from 'models/organization/organization';
 import { OutgoingWebhookStore } from 'models/outgoing_webhook/outgoing_webhook';
 import { ResolutionNotesStore } from 'models/resolution_note/resolution_note';
@@ -32,6 +33,7 @@ import { UserGroupStore } from 'models/user_group/user_group';
 import { makeRequest } from 'network';
 import { AppFeature } from 'state/features';
 import PluginState from 'state/plugin';
+import { retryFailingPromises } from 'utils/async';
 import {
   APP_VERSION,
   CLOUD_VERSION_REGEX,
@@ -44,6 +46,9 @@ import FaroHelper from 'utils/faro';
 // ------ Dashboard ------ //
 
 export class RootBaseStore {
+  @observable
+  isBasicDataLoaded = false;
+
   @observable
   currentTimezone: Timezone = moment.tz.guess() as Timezone;
 
@@ -82,7 +87,7 @@ export class RootBaseStore {
   @observable
   onCallApiUrl: string;
 
-  // --------------------------
+  // stores
   userStore = new UserStore(this);
   cloudStore = new CloudStore(this);
   directPagingStore = new DirectPagingStore(this);
@@ -105,10 +110,10 @@ export class RootBaseStore {
   globalSettingStore = new GlobalSettingStore(this);
   filtersStore = new FiltersStore(this);
   labelsStore = new LabelStore(this);
+  loaderStore = LoaderStore;
 
-  // stores
-
-  async updateBasicData() {
+  @action.bound
+  async loadBasicData() {
     const updateFeatures = async () => {
       await this.updateFeatures();
 
@@ -119,18 +124,21 @@ export class RootBaseStore {
       }
     };
 
-    return Promise.all([
-      this.userStore.loadCurrentUser(),
-      this.organizationStore.loadCurrentOrganization(),
-      this.grafanaTeamStore.updateItems(),
-      updateFeatures(),
+    await retryFailingPromises([
+      () => this.userStore.loadCurrentUser(),
+      () => this.organizationStore.loadCurrentOrganization(),
+      () => this.grafanaTeamStore.updateItems(),
+      () => updateFeatures(),
+    ]);
+    this.isBasicDataLoaded = true;
+  }
+
+  @action.bound
+  async loadMasterData() {
+    Promise.all([
       this.userStore.updateNotificationPolicyOptions(),
       this.userStore.updateNotifyByOptions(),
       this.alertReceiveChannelStore.updateAlertReceiveChannelOptions(),
-      this.outgoingWebhookStore.updateOutgoingWebhookPresets(),
-      this.escalationPolicyStore.updateWebEscalationPolicyOptions(),
-      this.escalationPolicyStore.updateEscalationPolicyOptions(),
-      this.escalationPolicyStore.updateNumMinutesInWindowOptions(),
     ]);
   }
 
@@ -280,7 +288,7 @@ export class RootBaseStore {
     return this.license === GRAFANA_LICENSE_OSS;
   }
 
-  @observable
+  @action.bound
   async updateFeatures() {
     const response = await makeRequest('/features/', {});
     this.features = response.reduce(